requirements.txt
aiohttp-sse-client==0.2.1
finta==1.3
PyYAML==6.0.2
requests==2.32.4
pandas==2.3.0
aiohttp==3.12.13
matplotlib==3.10.3